requirements.txt
flask
flask-cors
tensorflow
opencv-python-headless
numpy
pillow
werkzeug
